added auth hook
